requirements.txt
aiogram==3.22.0
APScheduler==3.11.1
python-dotenv==1.2.1
fastapi[standard]==0.115.0
httpx==0.28.1
perplexityai==0.20.0
pydantic==2.11.10
sqlmodel==0.0.27

